feat : api 명칭 정리 및 수정
